requirements.txt
requests==2.26.0
xlsxWriter==3.1.2
msgraph-sdk==1.0.0a13
